refactor(server): extract IntentRoute to fix TS2589 type depth error
Follow existing pattern (ProjectRoute, TuiRoute) by extracting intent
endpoints to separate Hono instance. Fixes TypeScript type chain
depth limit exceeded error.

API change: /session/:sessionID/intent/:intentID -> /intent/:sessionID/:intentID
